ldlinux: Don't alloc initramfs unless initrd= is specified
It only makes sense to allocate memory for 'initramfs' if an initrd=
argument was on the command line, otherwise we're wasting memory.

Furthermore needlessly allocating 'initramfs' actually caused old
kernels, such as memtest86, to fail to run because of the following
check in syslinux_boot_kernel(),

    if (initramfs && hdr.version < 0x0200)
        goto bail;              /* initrd/initramfs not supported */

Move the allocation after we've checked for and found an initrd=
argument.

Signed-off-by: Matt Fleming <matt.fleming@intel.com>