mk: Don't link against compiler's version of libgcc
Syslinux provides its own copy of libgcc functions in
com32/lib/libgcc, so we don't need to link against the version
provided by the compiler. What's worse, the behaviour is different
between the two versions.

This caused problems, for example, when parsing the TOTALTIMEOUT
config directive in com32/menu because of the following calculation,

	totaltimeout = (atoll(skipspace(p + 13)) * CLK_TCK + 9) / 10;

which invokes __divdi3() as provided by libgcc. The results returned
by the Syslinux and compiler's copy of libgcc differ, leading to
differences in behaviour.

Signed-off-by: Matt Fleming <matt.fleming@intel.com>