ldlinux: Make the default execute type KT_KERNEL
Default to trying to boot a linux kernel image, which was the
behaviour of the old asm image file loader.

Note that we don't need to go through __intcall() to load a kernel as
the dynamic linker will fixup any references to the kernel load
functions for us at runtime. Furthermore, modules invoking execute()
could be located anywhere in the 32-bit address space and may not be
able to represent their pointer arguments with SEG() and OFFS(), so
using the comboot API may not even work.

Signed-off-by: Matt Fleming <matt.fleming@intel.com>