core: Partial revert of commit 9333426b and unify lmalloc()
The reason behind commit 9333426b ("elflink: fix the global naming for
lmalloc") seems to be that we need to avoid having two 'lmalloc'
symbols, one in the core and one in the com32 library.

Unfortunately, this commit introduced the following warning in
multiple places,

meminfo.c:47:2: warning: implicit declaration of function ‘lmalloc’
meminfo.c:47:9: warning: assignment makes pointer from integer without a cast
meminfo.c:93:5: warning: implicit declaration of function ‘free’
meminfo.c:93:5: warning: incompatible implicit declaration of built-in function ‘free’

Consolidate the implementations of lmalloc() so that it's suitable for
use both in the com32 library code and the core.

Signed-off-by: Matt Fleming <matt.fleming@intel.com>