ldlinux: Fix check for last command line entry
The expression that checks to see when we've reached the end of the
command line entries is incorrect, probably because the list macros
weren't used. Use list_is_last() to make the expression correct and
self-documenting.

This bug was highlighted by the following compiler warning,

cli.c: In function ‘cmd_reverse_search’:
cli.c:102:20: warning: comparison of distinct pointer types lacks a cast

Signed-off-by: Matt Fleming <matt.fleming@intel.com>