mbr: Add support for loading VBRs from XFS filesystems
Before reading the VBR into 0x7C00, we need to check the first 4 bytes of
the read sector and check whether it is a XFS superblock or not. If it
is, then we need to read the right sector, which is the 4th sector of
the XFS partition where the Syslinux boot sector lives in.

This patch adds support for booting off XFS partitions properly on
MBR HDDs.

Signed-off-by: Paulo Alcantara <pcacjr@zytor.com>