xfs: Initial skeleton for XFS filesystem support
And another filesystem driver for Syslinux. Heh :-)

This patch *only* prints out a sample message on the screen when it
finds an active partition formated as a XFS filesystem. If this really
happens, then the XFS filesystem support in the EXTLINUX installer is
working nicely and it's time to write the driver.

Signed-off-by: Paulo Alcantara <pcacjr@zytor.com>