xfs: Implement xfs_fs_init() function
The XFS filesystem driver is now able to read properly the XFS superblock
and store useful information to be used around the driver code, which is
kept (in memory) in the xfs_fs_info structure.

As XFS filesytems store their superblock's fields in big-endian format,
we need to have functions capable to transform a 16/32/64-bit big-endian
value into a 16/32/64-bit little-endian one, since Syslinux is only able
to boot up on x86 systems, which store values in litte-endian format. So,
the bexx_to_cpu() functions do this trick.

The driver isn't functional yet, but I'm looking forward to implement
the iget() family of functions ASAP, then we can to at least get a
"boot:" from a XFS filesystem.

Signed-off-by: Paulo Alcantara <pcacjr@zytor.com>