Architecture dependency in LZO data compression library in subdirectory core/lzo/ is taken care to be able to build fo i386 for x86_64.
The implementation of eflags() in call16.c supported only 32bit, but now includes x86_64
as well. com32 lib build depends on it.

The file comboot.inc needed alignment specification for the variable ConfigName
in order to keep the linker happy. It's not clear why this would be needed for EFI.

Remanants of the unused old i386-only files, if any, need to be pruned.