This is the first cut checkin for GOP supporting the menu subsystem. For the initial cut, all of vesa code is dropped into vesa/efi with code buildable for EFI with a build option. There is VESA only code that is in the efi subdirectory as well that should go away eventually. Other files not dependent on EFI or not relevant to EFI need to be removed as well (e.g. i915resolution.c).
The crux of the code is in initvesa.c and screencpy.c. The GOP
initialization code sets up a matching graphics mode for the caller's
resolution needs. The VESA data structure is populated and mapped to
the EFI implementation as much as possible with some EFI extensions.
Since there is no way to query EFI for font, raw font data for
cp865_8x16 is added.

The code should build. Initial issues with testing the code are:
GOP mode on qemu-x86_64 does not match 640x480 and secondly
querying for the mode info on qemu fails. However, a simple program
querying GOP modes on efi64 platform works ok though. Testing the
resulting syslinux.efi on an efi64 platform does not seem to print output
other than in efi_main() routine . I suspect that it may be due to loading
failure or file system initialization. The issue does not show up on
qemu. In summary, there are issues on qemu and an efi64 platform that
need to be resolved.