xfs: rework the logic of xfs_get_ino_core()
(Just remove a few lines of debugging dirty marks. Sorry for
 the previous carelessness.)

Since we can calculate the real blk address directly by a given
ino, there is no need for implementing such complex logic which
would look up an inode from agi_root.

It seems that XFS put 64 inodes into one logic chunk for
performance consideration. That is to say, if our purporse is
just to implement a workable read-only xfs support, there is
also no need to handle this case, which do reduce lots of
complexity.

Also, in order to handle the inode whose starting address is
in the middle of one block, we need to hold the info of
i_block_offset, which can be calculated by XFS_INO_TO_OFFSET.

Signed-off-by: Chen Baozi <baozich@gmail.com>
Signed-off-by: Paulo Alcantara <pcacjr@zytor.com>