xfs: Fix memory leaks in xfs_dir2_node_find_entry() function
This function has been leaking memory when it found or not an entry,
since the leaf pointer were not being freed on either sucesss or fail.

Also, after doing a binary search to find the hash we're interested in,
and if it wasn't found, then the entry we want is not in that current
directory, so we do not need to print out an error message in this case.

Signed-off-by: Paulo Alcantara <pcacjr@zytor.com>