xfs: Add xfs_fmt_local_find_entry() function
Do not call xfs_dir2_local_find_entry() directly. As there is a
xfs_fmt_extents_find_entry() which lookups entries in inodes (with format
"extents") in a generic way, we can have a xfs_fmt_local_find_entry()
function which lookups entries in inodes (with format "local") in a
generic way too.

Signed-off-by: Paulo Alcantara <pcacjr@zytor.com>