xfs: Add xfs_fmt_local_readdir() function
Do not call xfs_fmt_local_readdir() directly. As there is a
xfs_fmt_extents_readdir() function which lists directory entries in
inodes (with format "extents") in a generic way, we can have a
xfs_fmt_local_readdir() function which lists directory entries in inodes
(with format "local") in a generic way too.

Signed-off-by: Paulo Alcantara <pcacjr@zytor.com>