EXTLINUX: Add sanity check for XFS filesystems
Syslinux won't work on >4 KiB filesystem block sizes since there is no
left space in MBR to determine where to install Syslinux bootsector.

If one is trying to install Syslinux in a XFS partition with a
filesystem block size different of 4 KiB, a proper error will be
returned to the user informing that the current filesystem block size
isn't supported by Syslinux.

Reported-by: Gene Cumm <gene.cumm@gmail.com>
Signed-off-by: Paulo Alcantara <pcacjr@zytor.com>