core: Change numIPAppends data type
For EFI we don't have the opportunity to mess with the object type of
'numIPAppends' via assembly as is done for BIOS - we need to do it all
in C. So change numIPAppends to size_t to accurately reflect its use.

Without this change the loop in __syslinux_get_ipappend_strings()
actually uses the address of numIPAppends instead of any value stored
there because it's a char array.

Signed-off-by: Matt Fleming <matt.fleming@intel.com>