efi, ansi: Improve console handling even further (e.g. scrolling support)
There were quite a few trivial bugs in the EFI console code, such as
mixing up column and row arguments and implementing functions with
different semantics when compared with the BIOS code.

With these changes console output now looks the same as the BIOS
version, with output scrolling down the screen instead of overwriting
existing text at the top whenever it incremented past the last row.

Signed-off-by: Matt Fleming <matt.fleming@intel.com>