xfs: Add full B+tree search support in xfs_dir2_node_find_entry()
According to the manual, the difference between B+tree directory
and node directory is that the node/leaf trees can be more than
one level deep in the B+tree format. That is to say we do not
need to reimplemented a completely new function in B+tree case.
What we've done here is to modify xfs_dir2_node_find_entry() to
support searching in a B+tree more than one level.

Signed-off-by: Chen Baozi <baozich@gmail.com>
Signed-off-by: Paulo Alcantara <pcacjr@zytor.com>