xfs: Add XFS_DINODE_FMT_BTREE support in xfs_next_extent()
XFS's extent B+Tree is organized in the key of offset. So we cannot
traverse the tree to look up the given extent number. Luckily, the
tree has been already threaded. So what we have to do is just to find
the first extent’s leaf, then traverse in a "linked list".

Signed-off-by: Chen Baozi <baozich@gmail.com>
Signed-off-by: Paulo Alcantara <pcacjr@zytor.com>