efi, vesa: Fix vesa initialization and video mode resolution
Fixed undefined symbol dprintf for production build. Now dprintf.h is
included so that both debug and production builds are ok.

Fixed the initialization of mode_info structure to point to a valid
GOP mode structure prior to printing it.

Fixed the case when the requested video mode resolution is unsupported
on a platform.  Essentially, VESA initialization is passed a desirable
video resolution. The underlying platform may either grant it or
return a system supported default value.  This fix is necessitated by
EFI-enabled platforms. Appropriate changes to callers of the vesa
initialization and the BIOS centric code are made as well.

Signed-off-by: Chandramouli Narayanan <chandramouli.narayanan@intel.com>
Signed-off-by: Matt Fleming <matt.fleming@intel.com>