kontron_wdt: Adding watchdog com32 module
When using a Kontron ETX board, it's possible to initialize and start
the watchdog during syslinux booting.

This allow protecting a boot sequence with a defined timeout.

Bootloader is starting, engage the watchdog and then start a default
entry (typically a Linux image).

If the loaded OS, feed or reinitalize the watchdog, nothing occurs
unless the system will reboot

Conflicts:

	com32/modules/Makefile