efi: We don't need two memory_map() functions
There are two functions for getting the EFI memory map, each with a
different sets of semantics. get_memory_map() returns the number of
entries in the map in it's first parameter, whereas memory_map()
returns the size of the memory map. The implementation of memory_map()
is bogus as it mixes up the semantics by calling get_memory_map().

The two functions do pretty much the same thing anyway so delete the
buggy one.

Signed-off-by: Matt Fleming <matt.fleming@intel.com>