ldlinux: Use stdcon instead of rawcon for the console
Switch to the stdcon input device so things like fgets() work
correctly when reading from stdin.

I noticed this was broken when running dmitest.c32. fgets() was
returning NULL because there was no pending user input instead of
waiting for/buffering the input. Modules are expecting stdin to be a
line-oriented device, not a raw one where each module has to implement
its own line support.

Signed-off-by: Matt Fleming <matt.fleming@intel.com>