linux.c32: Introduce initrdfile= option
It is useful to be able to load a file and pass it into a kernel's
rootfs via the initramfs scheme.  Given "initrdfile=foo", we will
load the file foo, encapsulate it with the initramfs cpio format,
then pass it alongside any initramfs files that were specified by
"initrd=" and "initrd+=" options.

One can specify the desired path/filename for the file to have
within the rootfs by using the at (@) sign, as in:

  initrdfile=foo@/goes/to/foo

One can also specify multiple files, separated by commas, such as:

  initrdfile=foo,bar@/somewhere/bar,baz

One can also use this option multiple times, as in:

  initrdfile=foo,bar initrdfile=baz@/somewhere/baz

Signed-off-by: Shao Miller <sha0.miller@gmail.com>