ldlinux: Fix logic if no DEFAULT or UI directive is found
Somewhere along the way the code that prints,

    No DEFAULT or UI configuration directive found!

was broken, and now no longer prints at all. While we're fixing this
it's a good opportunity to rework the logic to be clearer. Now we only
print the message if a config file was found, since these directives
are obviously missing if there is no config file (a warning will be
printed about the lack of config file anyway).

Signed-off-by: Matt Fleming <matt.fleming@intel.com>