core: Fix realloc() code transcription error
Commit 79459f631546eea83d4158f535c20ebd4ac18987 copied portions of
com32/lib/realloc.c into core/mem/malloc.c, with minor changes
for accessing allocation arena header attributes.  The previous
code used structure members and the current code uses bitmask
macros.

On a particular line in the original realloc(), there were two
assignments in a single expression that included a compound
assignment.  This fact was missed when the code was copied and
modified to use the bitmask macros.

Signed-off-by: Shao Miller <sha0.miller@gmail.com>
Signed-off-by: Matt Fleming <matt.fleming@intel.com>