mk: Add Makefile variable for python
Instead of hard-coding the python executable name use a make variable
like we do for all other executables used during the build. This has
the added bonus of allowing $(PYTHON) to be overridden on the command
line, e.g.

   make PYTHON=python2

which is particularly useful for distributions where /usr/bin/python
is actually python3.

Signed-off-by: Matt Fleming <matt.fleming@intel.com>