Makefile: Move Makefile.private inclusion to change default target
If no make target is specified on the command line we'll try to build
the first target we encounter in Makefile. Because we include
Makefile.private before any targets that means we'll try to build
isolinux.iso by default, which probably isn't what the user expected.

Move Makefile.private *after* our first target in Makefile so that the
user needs to explicitly build the targets in Makefile.private.

Signed-off-by: Matt Fleming <matt.fleming@intel.com>