efi: Consume unprocessed user input before loading ldlinux.c32
Consume and discard any pending user input that is received before we
load ldlinux.c32 and present the user with a "boot:" prompt.

This bug was discovered after commit 7cb503fd746b ("ldlinux: get_key()
requires raw access to user input") because user input was previously
handled with __stdcon_read() which would discard any incoming newline
characters. It turns out that if syslinux.efi is invoked from the EFI
shell the shell doesn't process the last newline character generated
when the user hits the return key and __raw_read() was returning this
as valid input.

Signed-off-by: Matt Fleming <matt.fleming@intel.com>