efi: Add -mno-red-zone to CFLAGS
The 'red zone' is a 128-byte area at the bottom of the stack frame and
is defined as part of the SysV Application Binary Interface. However,
UEFI firmware uses the Microsoft ABI, rather than the SysV ABI, and
our red zone can become corrupt if we don't turn it off.

Signed-off-by: Matt Fleming <matt.fleming@intel.com>