CLI: Delete 'temp_cmdline' and put 'cmdline' on the stack
We don't need to preserve the contents of 'cmdline' across calls to
edit_cmdline(). In fact, doing so causes things like the TAB key to
not function properly because the previous command is used as a prefix
to print_labels(), as though the user had typed it explicitly before
hitting TAB.

Delete 'temp_cmdline' because it no longer serves a purpose.

Signed-off-by: Matt Fleming <matt.fleming@intel.com>